requirements.txt
streamlit
sentence-transformers
transformers
torch
faiss-cpu
pdfplumber
python-docx
numpy
pandas
